requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
line-bot-sdk==3.11.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.31.0
pandas==2.2.2
weaviate-client
openai
uvicorn[standard]
tqdm
python-docx  # 若需要產出 Word

